Move the inlining code into its own module.
